+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 0.0.2
+
+- Added a new masked Personality (Insane) - may rename this later, but for now this is what it is called, the Insane will enter randomised signal translator messages to confuse/help/kill you. More functionality will come soon.
+- Fixed the "Deceiving" Personality where in it seemed like it didnt actually do anything at the terminal, it will now enter terminal codes at the terminal for a period of time (a set amount of codes), there is a randomised time delay between codes.
+
 ### 0.0.1
 
-- Initial release, a 99% replica of Piggy's MaskedAIRevamp (Unstable), this will be uploaded to thunderstore just before 0.0.2 release and is more going to be there for reference to the original base which this mod started from.
+- Initial release version number, there is no actual release, if you want this version, please head over to Piggy's MaskedAIRevamp and use version 0.2.1 -> https://thunderstore.io/c/lethal-company/p/Piggy/MaskedAIRevamp/.